Remove react
The primary two components (LaunchBinder & Shell) had 0 dom
interactions, once a term instance was provided to them. Term
being a component was also big overkill - it is already a class,
and we aren't doing anything extra here at all.

The Router was a good reason, so we'll have to do something
small that can be used as a replacement. But, it shouldn't be
too hard.

The experiment was useful and a success, I would think