Refactor to use a small router component
Didn't want to use any of the frameworks, since this feels
like such a small change. We needed to run cleanup functions
when the view changes - to disconnect websockets, cleanup terminals,
etc. This router gives us that.

I'll be kinda fucked if / when we re-introduce react or somesuch
here lol.